+++ llvm/utils/lit/lit/llvm/config.py
@@ -328,6 +328,9 @@ class LLVMConfig(object):
         return False
 
     def make_itanium_abi_triple(self, triple):
+        if triple == "wasm32-wasi":
+            return triple
+
         m = re.match(r"(\w+)-(\w+)-(\w+)", triple)
         if not m:
             self.lit_config.fatal(
@@ -339,6 +342,9 @@ class LLVMConfig(object):
         return m.group(1) + "-" + m.group(2) + "-" + m.group(3) + "-gnu"
 
     def make_msabi_triple(self, triple):
+        if triple == "wasm32-wasi":
+            return triple
+
         m = re.match(r"(\w+)-(\w+)-(\w+)", triple)
         if not m:
             self.lit_config.fatal("Could not turn '%s' into MS ABI triple" % triple)